+++ src/test/scala/lns/scenes/game/enemy/FiresContinuouslyTest.scala
@@ -26,7 +26,7 @@ case class MyFiringModel(
     stats: Stats,
     status: Queue[EnemyStatus] = Queue((EnemyState.Attacking, 0)),
     val crossable: Boolean = false,
-    life:Int = 0,
+    life:Double = 0,
     invincibilityTimer:Double = 0,
     fireRateTimer: Double = 0,
     shot: Option[Vector2] = None
@@ -39,7 +39,7 @@ case class MyFiringModel(
 
   def withStats(stats: Stats): Model                                 = copyMacro
   def withStatus(status: Queue[EnemyStatus] ): Model                        = copyMacro
-  def withAlive(life: Int, invincibilityTimer: Double): Model      = copyMacro
+  def withAlive(life: Double, invincibilityTimer: Double): Model      = copyMacro
   def withFire(fireRateTimer: Double, shot: Option[Vector2]): Model = copyMacro
 }
